Remove unused Export ingest gRPC call
