+++ nucliadb_protos/writer.proto
@@ -457,10 +457,6 @@ message ResourceIdResponse {
     string uuid = 1;
 }
 
-message ExportRequest {
-    string kbid = 1;
-}
-
 message SetVectorsRequest {
     utils.VectorObject vectors = 1;
     string kbid = 2;
@@ -561,7 +557,6 @@ service Writer {
     rpc Index(IndexResource) returns (IndexStatus) {}
     rpc ReIndex(IndexResource) returns (IndexStatus) {}
 
-    rpc Export(ExportRequest) returns (stream BrokerMessage) {}
     rpc DownloadFile(FileRequest) returns (stream BinaryData) {}
     rpc UploadFile(stream UploadBinaryData) returns (FileUploaded) {}
 }
